journal-repository: refactor
