+++ index.ts
@@ -1,5 +1,5 @@
 import { CleoAPI } from "./src/cleo-api/cleo-api";
-import { JournalsFileRepository } from "./src/journals-repository/journal-repository";
+import { JournalsFileRepository } from "./src/journal-repository/journal-repository";
 
 const PORT = 5011;
 const api = new CleoAPI(PORT, new JournalsFileRepository());
