+++ src/journal-repository/journal-entry-map.ts
