+++ src/journal-repository/journal-entry.ts
